move mocked_ifsg into dedicated pytest plugin
